requirements.txt
flask
openai-agents
waitress
pywebview
msal
bcrypt
pyttsx3
dotenv
elevenlabs
nltk
flask-cors
firebase-admin
google-cloud-texttospeech
python-dateutil
pyaudio
numpy
tensorflow
pygame